--- Assets/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e13efd1c4ba0b4693ae4d6dc37f4e678
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
