--- Assets/Editor/PatrolPathEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 574b9f9d90f1a4e0180d882b20207004
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
