+++ Assets/NewAudioMixer.mixer
@@ -1,0 +1,140 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!244 &-6688548534223334943
+AudioMixerEffectController:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: 
+  m_EffectID: cdfeb6900b7b1fd4dabd4ccab6211257
+  m_EffectName: Attenuation
+  m_MixLevel: 58ee59d379be1bf43bcf0303edc2bb0b
+  m_Parameters: []
+  m_SendTarget: {fileID: 0}
+  m_EnableWetMix: 0
+  m_Bypass: 0
+--- !u!244 &-1506531454791084207
+AudioMixerEffectController:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: 
+  m_EffectID: 671a47fb337c54e4bae245a594d612af
+  m_EffectName: Attenuation
+  m_MixLevel: 68c52afd05fed984dbea29110dcc3e13
+  m_Parameters: []
+  m_SendTarget: {fileID: 0}
+  m_EnableWetMix: 0
+  m_Bypass: 0
+--- !u!241 &24100000
+AudioMixerController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: NewAudioMixer
+  m_OutputGroup: {fileID: 0}
+  m_MasterGroup: {fileID: 24300002}
+  m_Snapshots:
+  - {fileID: 24500006}
+  m_StartSnapshot: {fileID: 24500006}
+  m_SuspendThreshold: -80
+  m_EnableSuspend: 1
+  m_UpdateMode: 0
+  m_ExposedParameters: []
+  m_AudioMixerGroupViews:
+  - guids:
+    - a45b6da85a8196f43acb06c2c00c09fe
+    - b7fc378a4e8649e41a2818b9bc8a41d0
+    - c7e4458db84adca4dba86eecac61ec61
+    name: View
+  m_CurrentViewIndex: 0
+  m_TargetSnapshot: {fileID: 24500006}
+--- !u!243 &24300002
+AudioMixerGroupController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Master
+  m_AudioMixer: {fileID: 24100000}
+  m_GroupID: a45b6da85a8196f43acb06c2c00c09fe
+  m_Children:
+  - {fileID: 8614813422439789519}
+  - {fileID: 5716237345009130210}
+  m_Volume: ae5e86c1e01f0214b9f3c083041a7cf0
+  m_Pitch: 96d4f00019c8b1c469254e8bee6abdbc
+  m_Send: 00000000000000000000000000000000
+  m_Effects:
+  - {fileID: 24400004}
+  m_UserColorIndex: 0
+  m_Mute: 0
+  m_Solo: 0
+  m_BypassEffects: 0
+--- !u!244 &24400004
+AudioMixerEffectController:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: 
+  m_EffectID: dee399f272299ad4980a5113f04af7ce
+  m_EffectName: Attenuation
+  m_MixLevel: 5ef8fdeb9c739bd44a5fda0f09b79845
+  m_Parameters: []
+  m_SendTarget: {fileID: 0}
+  m_EnableWetMix: 0
+  m_Bypass: 0
+--- !u!245 &24500006
+AudioMixerSnapshotController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Snapshot
+  m_AudioMixer: {fileID: 24100000}
+  m_SnapshotID: 8110e2295034d594db3917587d5faad0
+  m_FloatValues:
+    ae5e86c1e01f0214b9f3c083041a7cf0: -0.02577656
+    3b7cac3f2685d434891137633121f06a: 20
+  m_TransitionOverrides: {}
+--- !u!243 &5716237345009130210
+AudioMixerGroupController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Title music
+  m_AudioMixer: {fileID: 24100000}
+  m_GroupID: c7e4458db84adca4dba86eecac61ec61
+  m_Children: []
+  m_Volume: d67e7c3324d42984aa3d21bd3135af38
+  m_Pitch: 59992dc20ee9ed048a5be1f94d26baa8
+  m_Send: 00000000000000000000000000000000
+  m_Effects:
+  - {fileID: -1506531454791084207}
+  m_UserColorIndex: 0
+  m_Mute: 0
+  m_Solo: 0
+  m_BypassEffects: 0
+--- !u!243 &8614813422439789519
+AudioMixerGroupController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Game Music
+  m_AudioMixer: {fileID: 24100000}
+  m_GroupID: b7fc378a4e8649e41a2818b9bc8a41d0
+  m_Children: []
+  m_Volume: 3b7cac3f2685d434891137633121f06a
+  m_Pitch: 263d6a8de933041489bdf0c60c92846c
+  m_Send: 00000000000000000000000000000000
+  m_Effects:
+  - {fileID: -6688548534223334943}
+  m_UserColorIndex: 0
+  m_Mute: 0
+  m_Solo: 0
+  m_BypassEffects: 0
