+++ Assets/NewAudioMixer.mixer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 47720ed0024c83944b8db1ef543e5ec9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 24100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
