+++ Assets/Remnants/Ruby Gameplay/Death.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Death
+   : MonoBehaviour
+{
+
+    public Transform respawnPoint;
+
+
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+        if (collision.CompareTag("Death"))
+        {
+            transform.position = respawnPoint.position;
+        }
+      
+      
+    }
+}
