+++ Assets/Remnants/Ruby Gameplay/Death.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93785fc06e5b3bf4f8d51f655843b0d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
