+++ Assets/Remnants/Ruby Gameplay/GameOver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1443e732056122946b34d0835b6bae7b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
