+++ Assets/Remnants/Scenes/Credits.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d75475d9c9636ca43891ce71303973f0
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
