+++ UserSettings/EditorUserSettings.asset
@@ -21,10 +21,13 @@ EditorUserSettings:
       value: 5606565603005c0a5c0d597114750a4445154d7d2f7c7f32287b4935b7b4643d
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 56030c5654030b0c085f0f7615275c4412154a2c2e7b70362f284b67b7b3376c
+      value: 0201005f55075a03555c5d7313205d444f4e4b2e292b746174791831b3b96669
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 0201005f55075a03555c5d7313205d444f4e4b2e292b746174791831b3b96669
+      value: 56030c5654030b0c085f0f7615275c4412154a2c2e7b70362f284b67b7b3376c
+      flags: 0
+    RecentlyUsedSceneGuid-7:
+      value: 0706015252000d020e575a7146205e44454f417b2f2d71617e794e6ab5b33268
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
